backend/requirements.txt
# FastAPI框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# LLM和RAG
langchain==0.1.4
langchain-community==0.0.16
openai==1.10.0
chromadb==0.4.22
sentence-transformers==2.3.1

# 文档处理
pypdf==4.0.1
python-docx==1.1.0
unstructured==0.12.0

# 工具库
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
tenacity==8.2.3

